refined parameter matching for NetModels

+++ tests/test_particle_swarm.py
@@ -1,6 +1,6 @@
 from .context import data, algorithms, pset, objective, config, parse
 import numpy.testing as npt
-from os import mkdir, environ
+from os import mkdir, path
 from shutil import rmtree
 from copy import deepcopy
 
@@ -65,14 +65,17 @@ class TestParticleSwarm:
             ('random_var', 'v1'): [0, 10], ('random_var', 'v2'): [0, 10], ('random_var', 'v3'): [0, 10],
             'models': {'bngl_files/parabola.bngl'}, 'exp_data': {'bngl_files/par1.exp'},
             'bngl_files/parabola.bngl': ['bngl_files/par1.exp'], 'output_dir': 'test_pso_lh',
-            'initialization': 'lh', 'fit_type': 'pso',})
+            'initialization': 'lh', 'fit_type': 'pso'})
 
     @classmethod
     def teardown_class(cls):
         rmtree('test_pswarm_output')
-        rmtree('test_pso_lh')
-        rmtree('test_pso2')
-        rmtree('test_pso')
+        if path.isdir('test_pso_lh'):
+            rmtree('test_pso_lh')
+        if path.isdir('test_pso2'):
+            rmtree('test_pso2')
+        if path.isdir('test_pso'):
+            rmtree('test_pso')
 
     def test_random_pset(self):
         ps = algorithms.ParticleSwarm(deepcopy(self.config2))
@@ -134,18 +137,17 @@ class TestParticleSwarm:
                 assert ps.bests[i][0] in start_params
         assert count == 1
 
-    # def test_full(self):
-    #     conf_dict = parse.load_config(self.config_path)
-    #     myconfig = config.Configuration(conf_dict)
-    #     myconfig.config['bng_command'] = environ['BNGPATH'] + '/BNG2.pl'
-    #     ps = algorithms.ParticleSwarm(myconfig)
-    #     ps.run()
-    #     # print(ps.global_best)
-    #     best_fit = ps.global_best[0]
-    #
-    #     # The data is most sensitive to the x^2 coefficent, so this gets fit the best.
-    #     # Here's a reasonable test that the fitting went okay.
-    #     assert abs(best_fit['v1__FREE__'] - 0.5) < 0.3
+    def test_full(self):
+        conf_dict = parse.load_config(self.config_path)
+        myconfig = config.Configuration(conf_dict)
+        ps = algorithms.ParticleSwarm(myconfig)
+        ps.run()
+        # print(ps.global_best)
+        best_fit = ps.global_best[0]
+
+        # The data is most sensitive to the x^2 coefficent, so this gets fit the best.
+        # Here's a reasonable test that the fitting went okay.
+        assert abs(best_fit['v1__FREE__'] - 0.5) < 0.3
 
     def test_latin_hypercube(self):
         ps = algorithms.ParticleSwarm(self.lh_config)
